Update README.md
Running a Program with a Runtime Error

Run time errors—errors that do not occur until the program is run, and then may only occur with some data. These errors emphasize the need for completely testing your program.

Compile and run the program. Enter a number and press return. You should now see a run time error. There was no syntax or grammatical error in the program. However, just like commanding someone to break a law of nature, the program is asking the computer to break a law of math by dividing by zero. It cannot be done.

Correct this program by having the code divide by 2 instead of 0.

Re-compile and run the program. Type 9 when asked for input.

The runtime error should be gone, but does it gives the correct output? If not, change the data type of the number variable from int to float. Floating-point type variables are needed to hold numbers with a decimal point. Re-compile, run the program again, and see if entering 9 now gives the correct output.